requirements.txt
# Core Data Processing
pyspark==3.5.0
pandas==2.1.3
numpy==1.24.3

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Elasticsearch
elasticsearch==8.11.0

# Time Series Forecasting
prophet==1.1.5
statsmodels==0.14.0

# Machine Learning
scikit-learn==1.3.2

# Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
